+++ jsonpath/__init__.py
@@ -8,12 +8,19 @@ from .exceptions import JSONPathIndexError
 from .exceptions import JSONPathNameError
 from .exceptions import JSONPathSyntaxError
 from .exceptions import JSONPathTypeError
+from .exceptions import JSONPointerError
+from .exceptions import JSONPointerIndexError
+from .exceptions import JSONPointerKeyError
+from .exceptions import JSONPointerResolutionError
+from .exceptions import JSONPointerTypeError
 from .filter import UNDEFINED
 from .lex import Lexer
 from .match import JSONPathMatch
 from .parse import Parser
 from .path import CompoundJSONPath
 from .path import JSONPath
+from .pointer import JSONPointer
+from .pointer import resolve
 
 __all__ = (
     "compile",
@@ -30,8 +37,15 @@ __all__ = (
     "JSONPathNameError",
     "JSONPathSyntaxError",
     "JSONPathTypeError",
+    "JSONPointer",
+    "JSONPointerError",
+    "JSONPointerIndexError",
+    "JSONPointerKeyError",
+    "JSONPointerResolutionError",
+    "JSONPointerTypeError",
     "Lexer",
     "Parser",
+    "resolve",
     "UNDEFINED",
 )