+++ jsonpath/exceptions.py
@@ -76,6 +76,30 @@ class JSONPathNameError(JSONPathError):
         self.token = token
 
 
+class JSONPointerError(Exception):
+    """Base class for all JSON Pointer errors."""
+
+
+class JSONPointerEncodeError(JSONPointerError):
+    """An exception raised when a JSONPathMatch can't be encoded to a JSON Pointer."""
+
+
+class JSONPointerResolutionError(JSONPointerError):
+    """Base exception for those that can be raised during pointer resolution."""
+
+
+class JSONPointerIndexError(JSONPointerResolutionError, IndexError):
+    """An exception raised when an array index is out of range."""
+
+
+class JSONPointerKeyError(JSONPointerResolutionError, KeyError):
+    """An exception raised when a pointer references a mapping with a missing key."""
+
+
+class JSONPointerTypeError(JSONPointerResolutionError, TypeError):
+    """An exception raised when a pointer resolves a string against a sequence."""
+
+
 def _truncate_message(value: str, num: int, end: str = "...") -> str:
     if len(value) < num:
         return value